% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repec.R
\name{write_references}
\alias{write_references}
\title{Save references downloaded from RePEc IDEAS to a file}
\usage{
write_references(results, file_name)
}
\description{
Saves references retrieved using \code{\link{get_references}} to a file.
}
\seealso{
\code{\link{repec_search}} for performing the initial search, and
  \code{\link{get_references}} for retrieving reference data.
}
