% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Fetch full references for search results from RePEc IDEAS}
\usage{
get_references(results)
}
\arguments{
\item{results}{A data frame containing the URLs of one or more RePEc search
results (see \code{\link{repec_search}}).}
}
\value{
A new data frame, with reference data for each search result added.
}
\description{
Automatically fetches full reference data for a list of search results
returned by the \code{\link{repec_search}} function.
}
\seealso{
\code{\link{repec_search}} for performing the initial search, and
  \code{\link{write_references}} for saving the references to a file.
}
