% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repec.R
\name{repec_search}
\alias{repec_search}
\title{Search the RePEc IDEAS database.}
\usage{
repec_search(
  query,
  from_page = 0,
  to_page = 1000,
  publication_type = "all",
  search_in = "whole_record"
)
}
\arguments{
\item{query}{The search terms to use.}

\item{from_page}{Integer describing which page of results to start on (the first page is 0).}

\item{to_page}{Integer describing which page of results to end on.}

\item{publication_type}{String describing which publication type to search. Possible values: `all`, `articles`, `papers`, `chapters`, `books`, `software`}

\item{search_in}{String describing which fields to search in. Possible values: `whole_record`, `abstract`, `keywords`, `title`, `author`}
}
\value{
A data frame containing the URL for each search result.
}
\description{
`repec_search` performs a search and returns the results as a data frame.
}
